feat: 添加批处理执行摘要和进度报告功能，优化日志记录和数据集加载
