+++ .gitignore
@@ -13,7 +13,8 @@
 
 # 结果输出目录
 results/
-
+results_11_30/
+ 
 # 测试相关（如果您不想提交测试文件）
 tests/
